+++ controllers_v2/keysend.ctrl.go
@@ -8,7 +8,6 @@ import (
 	"strconv"
 
 	"github.com/getAlby/lndhub.go/common"
-	"github.com/getAlby/lndhub.go/db/models"
 	"github.com/getAlby/lndhub.go/lib/responses"
 	"github.com/getAlby/lndhub.go/lib/service"
 	"github.com/getAlby/lndhub.go/lnd"
@@ -72,7 +71,6 @@ type KeySendResponseBody struct {
 // @Security     OAuth2Password
 func (controller *KeySendController) KeySend(c echo.Context) error {
 	userID := c.Get("UserID").(int64)
-	limits := controller.svc.GetLimits(c)
 	reqBody := KeySendRequestBody{}
 	if err := c.Bind(&reqBody); err != nil {
 		c.Logger().Errorf("Failed to load keysend request body: %v", err)
@@ -83,7 +81,7 @@ func (controller *KeySendController) KeySend(c echo.Context) error {
 		c.Logger().Errorf("Invalid keysend request body: %v", err)
 		return c.JSON(http.StatusBadRequest, responses.BadArgumentsError)
 	}
-	errResp := controller.checkKeysendPaymentAllowed(context.Background(), reqBody.Amount, userID, limits)
+	errResp := controller.checkKeysendPaymentAllowed(c, reqBody.Amount, userID)
 	if errResp != nil {
 		c.Logger().Errorf("Failed to send keysend: %s", errResp.Message)
 		return c.JSON(errResp.HttpStatusCode, errResp)
@@ -110,7 +108,6 @@ func (controller *KeySendController) KeySend(c echo.Context) error {
 // @Security     OAuth2Password
 func (controller *KeySendController) MultiKeySend(c echo.Context) error {
 	userID := c.Get("UserID").(int64)
-	limits := controller.svc.GetLimits(c)
 	reqBody := MultiKeySendRequestBody{}
 	if err := c.Bind(&reqBody); err != nil {
 		c.Logger().Errorf("Failed to load keysend request body: %v", err)
@@ -130,7 +127,7 @@ func (controller *KeySendController) MultiKeySend(c echo.Context) error {
 	for _, keysend := range reqBody.Keysends {
 		totalAmount += keysend.Amount
 	}
-	errResp := controller.checkKeysendPaymentAllowed(context.Background(), totalAmount, userID, limits)
+	errResp := controller.checkKeysendPaymentAllowed(c, totalAmount, userID)
 	if errResp != nil {
 		c.Logger().Errorf("Failed to make keysend split payments: %s", errResp.Message)
 		return c.JSON(errResp.HttpStatusCode, errResp)
@@ -165,14 +162,14 @@ func (controller *KeySendController) MultiKeySend(c echo.Context) error {
 	return c.JSON(status, result)
 }
 
-func (controller *KeySendController) checkKeysendPaymentAllowed(ctx context.Context, amount, userID int64, limits *models.Limits) (resp *responses.ErrorResponse) {
+func (controller *KeySendController) checkKeysendPaymentAllowed(c echo.Context, amount, userID int64) (resp *responses.ErrorResponse) {
 	syntheticPayReq := &lnd.LNPayReq{
 		PayReq: &lnrpc.PayReq{
 			NumSatoshis: amount,
 		},
 		Keysend: true,
 	}
-	resp, err := controller.svc.CheckOutgoingPaymentAllowed(ctx, syntheticPayReq, userID, limits)
+	resp, err := controller.svc.CheckOutgoingPaymentAllowed(c, syntheticPayReq, userID)
 	if err != nil {
 		return &responses.GeneralServerError
 	}
